DBS-6195: Main job should be in warning if sub job has failed and DBS-6197: Installation of modules failed with error (#1899)
* DBS-6195 and DBS-6197

* DBS-6195 and DBS-6197

+++ server/src/operations/workloads/mssql/discover-consts.ts
@@ -827,8 +827,12 @@ const INSTALL_WF_POWERSHELL_PREREQS_PS1 = (requiredModules: string, s3SignedURL:
         if ($destinationPathExists -eq $False) {
             New-Item -ItemType Directory -Path $destinationPath -Force
         }
-    
-        Copy-Item "$Env:Temp\\dependent-packages\\powershell\\Microsoft.PackageManagement.NuGetProvider-2.8.5.208.dll" -Destination $destinationPath -Recurse -Force
+
+        $nugetDestinationPath = Join-Path -Path $destinationPath -ChildPath "Microsoft.PackageManagement.NuGetProvider-2.8.5.208.dll"
+        $nugetDestinationPathExists = Test-Path -Path $nugetDestinationPath
+        if ($nugetDestinationPathExists -eq $False) {
+            Copy-Item "$Env:Temp\\dependent-packages\\powershell\\Microsoft.PackageManagement.NuGetProvider-2.8.5.208.dll" -Destination $destinationPath -Recurse -Force
+        }
     
         $sourcelocation = "$Env:Temp\\dependent-packages\\aws"
         Import-PackageProvider -Name NuGet
